==> 01_scrape_fbref.ipynb <==
from scrape_utils import stats_to_parquet, VALID_STATS
import time
import random

def scrape_all(competition, years):
    for year in years:
        for stat in VALID_STATS:
            stats_to_parquet(competition, year, stat)
            time.sleep(random.randint(1, 10))

scrape_all('argentina1', [2022, 2023, 2024])

scrape_all('brazil1', [2022, 2023, 2024])

scrape_all('belgium1', [2022, 2023, 2024])

scrape_all('england1', [2022, 2023, 2024])

scrape_all('england2', [2022, 2023, 2024])

scrape_all('france1', [2022, 2023, 2024])

scrape_all('france2', [2022, 2023, 2024])

scrape_all('germany1', [2022, 2023, 2024])

scrape_all('germany2', [2022, 2023, 2024])

scrape_all('italy1', [2022, 2023, 2024])

scrape_all('italy2', [2022, 2023, 2024])

scrape_all('mexico1', [2022, 2023, 2024])

scrape_all('netherlands1', [2022, 2023, 2024])

scrape_all('portugal1', [2022, 2023, 2024])

scrape_all('spain1', [2022, 2023, 2024])

scrape_all('spain2', [2022, 2023, 2024])

scrape_all('usa1', [2022, 2023, 2024])
